refactor: Move standalone logic from main.rs to library crate
Move ~3100 lines of standalone client/server code from the binary
crate (main.rs) into two new library modules, following the existing
flat-file convention (benchmark.rs/benchmark_blocking.rs pattern).

Structure:
- standalone_server.rs (1982 lines): constants, shared helpers,
  server dispatch, multi-accept TCP/UDS, async server paths
- standalone_client.rs (1146 lines): retry helpers, client dispatch,
  single/concurrent blocking and async paths
- main.rs reduced from ~4200 to ~1120 lines (thin dispatch layer)

Additional changes:
- Promote logging.rs from binary-private to library-public module
- Move set_affinity() to utils.rs as pub function
- All standalone functions now pub for tarpaulin coverage measurement
  and integration test access

No behavioral changes. All 374 tests pass. Benchmark comparison
across 3 runs confirms no performance regression (mean latencies
within 2-5% run-to-run variance).
